Extract spell list and book into separate files
Moves the spell list/book UI from `control.cpp` code into separate files.
