+++ Source/diablo.cpp
@@ -37,8 +37,8 @@
 #include "error.h"
 #include "gamemenu.h"
 #include "gmenu.h"
-#include "hwcursor.hpp"
 #include "help.h"
+#include "hwcursor.hpp"
 #include "init.h"
 #include "lighting.h"
 #include "loadsave.h"
@@ -50,6 +50,7 @@
 #include "nthread.h"
 #include "objects.h"
 #include "options.h"
+#include "panels/spell_book.hpp"
 #include "pfile.h"
 #include "plrmsg.h"
 #include "qol/common.h"
